third_party/blink/web_tests/wpt_internal/html/canvas/drawElementImage/privacy/embed-image-ignored.https.sub.html

<!DOCTYPE HTML>
<html>
<head>
  <title>drawElementImage does not draw cross-origin embed element content</title>
  <script src='/resources/testharness.js'></script>
  <script src='/resources/testharnessreport.js'></script>
  <style>
  #child {
    width: 100px;
    height: 200px;
    background: green;
  }
  #sameOrigin {
    position: absolute;
    left: 0px;
    top: 0px;
  }
  #crossOrigin {
    position: absolute;
    left: 0px;
    top: 100px;
  }
  </style>
</head>
<body>
  <canvas id=canvas width="100" height="200" layoutsubtree>
    <div id=child>
      <embed id=sameOrigin width=100 height=100 type="image/png"
           src="https://{{location[host]}}/wpt_internal/html/canvas/drawElementImage/resources/red-100x100.png"/>
      <embed id=crossOrigin width=100 height=100 type="image/png"
           src="https://{{hosts[alt][www]}}:{{ports[h2][0]}}/wpt_internal/html/canvas/drawElementImage/resources/red-100x100.png"/>
    </div>
  </canvas>
  <script>
    window.onload = function() {
      test(function(t) {
        var context = canvas.getContext("2d");
        context.drawElementImage(child, 0, 0);

        pixel = context.getImageData(50, 50, 1, 1).data;
        assert_equals(pixel[0], 255, "Same origin should draw");
        assert_equals(pixel[1], 0, "Same origin should draw");
        assert_equals(pixel[2], 0, "Same origin should draw");

        pixel = context.getImageData(50, 150, 1, 1).data;
        assert_equals(pixel[0], 0, "Cross origin should not draw");
        assert_equals(pixel[1], 128, "Cross origin should not draw");
        assert_equals(pixel[2], 0, "Cross origin should not draw");

        t.done();
      });
    };
  </script>
</body>
</html>
